+++ packages/commerce-sdk-react/src/provider.tsx
@@ -22,6 +22,7 @@ import {
 } from 'commerce-sdk-isomorphic'
 import Auth from './auth'
 import {ApiClientConfigParams, ApiClients} from './hooks/types'
+import {Logger} from './types'
 
 export interface CommerceApiProviderProps extends ApiClientConfigParams {
     children: React.ReactNode
@@ -36,6 +37,7 @@ export interface CommerceApiProviderProps extends ApiClientConfigParams {
     enablePWAKitPrivateClient?: boolean
     clientSecret?: string
     silenceWarnings?: boolean
+    logger?: Logger
 }
 
 /**
@@ -75,6 +77,7 @@ export const AuthContext = React.createContext({} as Auth)
                     locale="en-US"
                     enablePWAKitPrivateClient={true}
                     currency="USD"
+                    logger={logger}
                 >
                     {children}
                 </CommerceApiProvider>
@@ -111,8 +114,13 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
         OCAPISessionsURL,
         enablePWAKitPrivateClient,
         clientSecret,
-        silenceWarnings
+        silenceWarnings,
+        logger
     } = props
+
+    // Set the logger based on provided configuration, or default to the console object if no logger is provided
+    const configLogger = logger || console
+
     const config = {
         proxy,
         headers,
@@ -127,6 +135,7 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
         throwOnBadResponse: true,
         fetchOptions
     }
+
     const apiClients = useMemo(() => {
         return {
             shopperBaskets: new ShopperBaskets(config),
@@ -167,7 +176,8 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
             OCAPISessionsURL,
             enablePWAKitPrivateClient,
             clientSecret,
-            silenceWarnings
+            silenceWarnings,
+            logger: configLogger
         })
     }, [
         clientId,
@@ -181,7 +191,8 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
         OCAPISessionsURL,
         enablePWAKitPrivateClient,
         clientSecret,
-        silenceWarnings
+        silenceWarnings,
+        configLogger
     ])
 
     // Initialize the session
@@ -200,7 +211,8 @@ const CommerceApiProvider = (props: CommerceApiProviderProps): ReactElement => {
                 shortCode,
                 locale,
                 currency,
-                silenceWarnings
+                silenceWarnings,
+                logger: configLogger
             }}
         >
             <CommerceApiContext.Provider value={apiClients}>