+++ packages/pwa-kit-react-sdk/src/utils/logger-instance.js
@@ -4,9 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+import createLogger from '@salesforce/pwa-kit-runtime/utils/logger-factory'
 
-const getShopperContext = ['organizationId', 'usid'] as const
-
-export default {
-    getShopperContext
-}
+const logger = Object.freeze(createLogger({packageName: 'pwa-kit-react-sdk'}))
+export default logger