+++ packages/template-retail-react-app/app/utils/logger-instance.js
@@ -4,9 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+import createLogger from '@salesforce/pwa-kit-runtime/utils/logger-factory'
 
-const getUrlMapping = ['organizationId', 'urlSegment', 'siteId', 'locale'] as const
-
-export default {
-    getUrlMapping
-}
+const logger = Object.freeze(createLogger({packageName: 'retail-react-app'}))
+export default logger